requirements.txt
# API Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# PDF Processing
pymupdf>=1.23.0

# Text Processing
unidecode>=1.3.8
regex>=2023.12.25

# Date/Time Utilities
python-dateutil>=2.8.2

# LLM Integration
openai>=1.12.0

# Optional: HTTP client (alternative to openai SDK)
httpx>=0.26.0

# Development & Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0

# Environment Variables
python-dotenv>=1.0.0
